refactor: add custom data source class
