+++ iOS/Airbnb/Airbnb/ViewControllers/ViewController.swift
@@ -12,6 +12,7 @@ class ViewController: UIViewController {
     @IBOutlet weak var collectionView: UICollectionView!
     
     private let searchViewController = SearchViewController()
+    private let dataSource = MainCollectionViewDataSource()
     
     private let searchBar: UISearchBar = {
         let searchBar = UISearchBar()
@@ -25,22 +26,24 @@ class ViewController: UIViewController {
         searchBar.delegate = self
         navigationItem.titleView = searchBar
         
-        collectionView.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "cell")
-        collectionView.register(RegionCollectionViewCell.self,
-                                forCellWithReuseIdentifier: RegionCollectionViewCell.reuseId)
-        collectionView.register(TravelStyleCollectionViewCell.self,
-                                forCellWithReuseIdentifier: TravelStyleCollectionViewCell.reuseId)
+        setupCollectionView()
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        tabBarController?.tabBar.isHidden = false
+    }
+    
+    private func setupCollectionView() {
+        collectionView.register(MainCollectionViewCell.self,
+                                forCellWithReuseIdentifier: MainCollectionViewCell.reuseId)
         collectionView.register(HeaderCollectionReusableView.self,
                                 forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
                                 withReuseIdentifier: HeaderCollectionReusableView.reuseId)
         collectionView.register(ImageHeaderCollectionReusableView.self,
                                 forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
                                 withReuseIdentifier: ImageHeaderCollectionReusableView.reuseId)
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        tabBarController?.tabBar.isHidden = false
+        collectionView.dataSource = dataSource
     }
 }
 
@@ -48,48 +51,6 @@ extension ViewController: UISearchBarDelegate {
     func searchBarTextDidBeginEditing(_ searchBar: UISearchBar) {
         searchBar.resignFirstResponder()
         navigationController?.pushViewController(searchViewController, animated: true)
-    }
-}
-
-extension ViewController: UICollectionViewDelegate {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 3
-    }
-}
-
-extension ViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cell", for: indexPath)
-        switch indexPath.section {
-        case 1:
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: RegionCollectionViewCell.reuseId, for: indexPath) as! RegionCollectionViewCell
-            return cell
-        case 2:
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TravelStyleCollectionViewCell.reuseId, for: indexPath) as! TravelStyleCollectionViewCell
-            return cell
-        default:
-            return cell
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: HeaderCollectionReusableView.reuseId, for: indexPath) as! HeaderCollectionReusableView
-        switch indexPath.section {
-        case 0:
-            let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: ImageHeaderCollectionReusableView.reuseId, for: indexPath) as! ImageHeaderCollectionReusableView
-            return header
-        case 1:
-            header.configure(title: "가까운 여행지 둘러보기")
-        case 2:
-            header.configure(title: "어디에서나 여행은 \n살아보는거야!")
-        default:
-            break
-        }
-        return header
     }
 }
